backend/requirements.txt
# Dependências do Backend - Sistema de Orações Igreja Videira
# EXCLUSIVAMENTE Supabase - Sem armazenamento local

# Framework web
fastapi==0.110.1
uvicorn==0.25.0

# Cliente Supabase - OBRIGATÓRIO
supabase>=2.4.0

# Validação de dados
pydantic>=2.6.4

# Utilitários essenciais
python-dotenv>=1.0.1
python-multipart>=0.0.9
requests>=2.31.0

# Data e tempo
tzdata>=2024.2

# Desenvolvimento e testes
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Removidas dependências desnecessárias para sistema exclusivo Supabase:
# - boto3 (AWS S3)
# - pymongo/motor (MongoDB)
# - requests-oauthlib (OAuth)
# - cryptography (criptografia avançada)
# - email-validator (validação de email)
# - pyjwt/python-jose (JWT tokens)
# - passlib (hash de senhas)
# - pandas/numpy (análise de dados)
# - jq/typer (CLI tools)
